:recycle: api: move CertificationAssessmentScoringV3 to src
