+++ api/lib/domain/events/handle-certification-rescoring.js
@@ -9,7 +9,7 @@ import { ChallengeDeneutralized } from './ChallengeDeneutralized.js';
 import { CertificationJuryDone } from './CertificationJuryDone.js';
 import { checkEventTypes } from './check-event-types.js';
 import { CertificationVersion } from '../../../src/shared/domain/models/CertificationVersion.js';
-import { CertificationAssessmentScoreV3 } from '../models/CertificationAssessmentScoreV3.js';
+import { CertificationAssessmentScoreV3 } from '../../../src/certification/scoring/domain/models/CertificationAssessmentScoreV3.js';
 import { ABORT_REASONS } from '../models/CertificationCourse.js';
 import { FlashAssessmentAlgorithm } from '../../../src/certification/flash-certification/domain/models/FlashAssessmentAlgorithm.js';
 import { config } from '../../../src/shared/config.js';
