+++ api/src/certification/scoring/domain/models/CertificationAssessmentScoreV3.js
@@ -1,6 +1,6 @@
-import { status as CertificationStatus } from '../../../src/shared/domain/models/AssessmentResult.js';
-import { config } from '../../../src/shared/config.js';
-import { ABORT_REASONS } from './CertificationCourse.js';
+import { status as CertificationStatus } from '../../../../shared/domain/models/AssessmentResult.js';
+import { config } from '../../../../shared/config.js';
+import { ABORT_REASONS } from '../../../../../lib/domain/models/CertificationCourse.js';
 
 const MINIMUM_ESTIMATED_LEVEL = -8;
 const MAXIMUM_ESTIMATED_LEVEL = 8;
